feat(backend-next): CORS with shared origin allowlisting
Moves isOriginTrusted and matchesWildcard into @c15t/schema and has
@c15t/backend re-export them, then wires CORS into backend-next.

Shared because it is a security decision: both backends answer for the same
tenants with the same trusted domains during the parallel phase, and an origin
accepted by one but rejected by the other is a hole or an outage depending on
direction. The matching is subtler than it looks — IPv6 literals carry their
port after the bracket, default ports are implicit per scheme, www. is
equivalent, and *. wildcards deliberately exclude the apex — so a second
implementation would drift on an edge case rather than obviously. The shipping
package's own 34 CORS tests pass against the shared version.

Three behaviours pinned in backend-next's tests. An untrusted origin gets no
CORS headers at all rather than an echoed origin, since echoing defeats the
allowlist. No configured origins rejects everything rather than defaulting
open. And a preflight is answered 204 before any route runs — it carries no
credentials by design, so requiring auth there would break every browser
client.

Vary: Origin is set on allowed responses so a shared cache cannot serve one
origin's response to another.

96 tests.
